CMS-11787 Move openDialog implementation to OpenUiService
Instead of letting each extension implement openDialog itself (by
invoking OpenUiService.openDialog with the right 'locals' for the
dialog), the OpenUiService can handle this itself. As a result, all
extensions (including future ones) will be able to open dialogs.

The exception is a dialog itself. Calls to openDialog from within a
dialog now return the error code 'DialogExists', which seems logical.

The result is in line with the Wicket implementation.