CMS-11787 Return 'DialogExists' error code when dialog already exists
when a dialog was opened from a dialog the first check on the dialogUrl
parameter was always triggered, while we should return an Error with
the code 'DialogExists'. This cannot be a true Error object since in
that case Penpal does not transfer the 'code' property correctly. The
trick is to pass an object with a 'message' and 'code' property, from
which Penpal will recreate a proper Error object.

The second, internal error (when dialogUrl does not exist) can simply
be thrown. Penpal will pass it to the calling promise.

Also added a message to the 'DialogCanceled' error.