CMS-11789 Refactor Validator interface
- remove getName()
- remove init(ValidatorContext) (moves into the validate() call,
  which can now throw a ValidationContextException if needed)

Also removed the AbstractFieldValidator base class, so we cannot put any
critical utility methods in there.

The functionality to create violations with translated error messages
moved into the ValidationContext class. Custom validators can either
use these methods, or return a violation as a lambda that directly
returns a non-translated message or a translated message not stored in
the repository.