CMS-11789 Provide access to translated violations
Using the ValidationContext to retrieve translated violations turned out
to be tricky: it required passing the current validator, and it required
a lookup of a validator's name based on a validator instance (which
would require something like ValidationService.getValidatorName(Validator)
and a BiMap to map names to validators and vice versa). That all looked
a bit ugly.

Fixed by introducing the ValidatorInstance class, which binds a
Validator to its ValidatorConfig. The ValidationService now returns
ValidatorInstances instead of Validators. Violations are now created
with a ViolationFactory, which is passed as the third argument to
Validator.validate(). A ValidatorInstance is a ViolationFactory: it
already knows the validator name from the ValidatorConfig, and keeps
track of the current locale in the ValidatorContext via a ThreadLocal.