CMS-11789 Introduce generic field value type in Validator API
Using a value type that does not match with the actual value of the
field at runtime will produce a ClassCastException in
ValidatorInstanceImpl.validate().

The CMS code cannot used these generic. The validator are created
without any context, so at that point the expected value type is
unknown. Hence the value type of ValidatorInstance is Object.

Both Wicket and the Channel Manager now pass the 'actual' Java object
to validators. In Wicket that's easy (just pass whatever is in the model
of the field). In the Channel Manager we need to convert the string
value we get from the client back to a corresponding Java type.