CMS-11789 Move creation of violations into validation context
The third 'violationFactory' argument to Validator.validate() felt too
verbose, and the repetition of the 'validate()' signature in Validator
and ValidatorInstance called for a shared base type.

With some type refactoring the violations can now be created via the
ValidationContext again, which looks good API-wise. The shared base type
'AbstractValidator' also makes it possible to reuse the definition of
the validate() method in both a Validator and ValidatorInstance. The
difference is the type of context they get: a ValidatorInstance only
gets a 'base' context that cannot lookup translations. It implements the
lookup of translations itself, and delegates all other ValidationContext
calls to the base context stored in the ThreadLocal. Each validator gets
a full context so it can lookup translations, which is implemented by
the ValidatorInstance that wraps it.