CMS-11789 Implement ValidationContext.getParentNode()
Renamed the BaseValidationContext class to FieldContext.
ValidationContext now is a FieldContext that can create violations.

To pass the parent node, the surrounding code has to create a specific
context per 'level' in the document. This is quite simple in Wicket.

In the content service, we pass a CompoundContext that contains the
node of the compound and the locale and time zone. Whenever the code
recurses into a compound or choice field a new compound context is
created with the correct child node.

The logic to execute validators has moved into AbstractFieldType, since
validators can be configured (and are executed) for *every* field (also
for compounds!). For all these compound fields the node of the compound
is passed as the validated value, so a Validator<Node> can be used to
validate it.