NAVAPP-10 Add basic left menu implementation
Basic left menu implementation without styles and the mock navigation configuration. Visual part is split up into several components which should nicely fit in the scope of styling issue but anyway it's a subject for possible future changes. Services mimic real behavior but the menu structure is not fully defined.

#resolve #assign EMPTY