CMS-11839 Load the navigation communication library
We do not deploy our javascript libraries to a CDN yet. So we need to
include them in the engine jar for now. This means we need to pull them
in via npm and then copy them to the target directory. Then we can use
a HeaderItem (NavAppHeaderItem) to generate a script source header in
the NavAppPage.html for it.

The NavAppHeaderItem will later also generate javascript for the
NavAppSettings object and generate a script header for the navigation
application angular app.