CHANNELMGR-2357 Pass correct context node to validators of node fields
LinkFieldTypes (ImageLinkFieldType and NodeLinkFieldType) used in
compounds did not get the right JCR node in their CompoundContext.
Instead of their own node, they got the node of their parent compound.

Fixed by moving the default implementation of the validate() method for
node fields to the NodeFieldType interface, so all node fields use it
automatically.

While refactoring this, I noticed that the readFrom(), readValues() and
writeValues() implementations were identical in all node fields, so
these could move to default methods in the NodeFieldType interface too.
The odd ones out are a) choice fields (which they are by nature, so
that's OK) and b) rich text fields (which should not inherit from
formatted text fields, but that's to be refactored later).

The default methods in NodeFieldType needed readValues() and
writeValues() in the parent interface instead of in AbstractFieldType.
The 'internal' methods validateValue() and getValidatedValue() moved
there too, so the 'external' API in FieldType becomes smaller.

LinkFieldType used to inherited from PrimitiveFieldType, which did not
make sense since they don't store their value in a property but in a
node. The new LeafFieldType captures all 'leaf' fields, i.e. fields that
don't contain nested fields. All primitive fields are leaf fields, but
so are the link fields. The rich text fields should become leaf fields
too.