CMS-11867 Add percentage height support to dialog content element
The height of a Wicket dialog is not controlled by the outer element of
a dialog, but by a deeply nested element (fortunately the width is set
on the outer element).

To still be able to set a percentage height on the dialog, we use the
fact that the percentage is always in context of the full viewport,
which means we can use the 'vh' unit instead of '%'. We do have to
adhere for the space the existing top and bottom borders occupy, as well
as the dialog header, so those are calculated and subtracted using CSS
calc().