CHANNELMGR-2357 Split RichTextFieldType and FormattedTextFieldType
A rich text field is a node field, yet it extended formatted text which
is a property field. As a result the rich text field had to explicitly
override a lot of parent methods to make sure the node field logic was
used instead.

The rich text and formatted text are now independent: rich text is now a
true node field. The shared logic of reading the CKEditor configuration
moved into an HtmlFieldConfig utility class. Both fields implement the
HtmlField interface to capture the fact that each of them should have
a 'config' property for the CKEditor configuration (which is expected
by the frontend).