CHANNELMGR-2354 Make resize handles invisible
The resize handle made the edge of the left side panel look odd: it cut
off all horizontal lines, except the line below the tabs which was
extended visually with a CSS hack.

Both resize handles are now invisible, and 'float' above the left or
right side of the right and left side nav, respectively. As a result
both size navs look cleaner and the styling got simpler.