CMS-11633 Add a 'testnamespace' fixture to the repository test content
Downstream projects of the repository *but* upstream of the cms project
require integration tests that require a test fixture below
hippo:namespaces. Therefor, add a 'testnamespace' fixture to the
test content that can be used by downstream projects like the content
type service and the hst. Note that the 'testnamespace' fixture does not
have or rely on hippo:editor or hippo:frontend like the namespace fixture
in the cms project does, since hippo:editor nad hippo:frontend are at
this moment namespaces provided by the CMS