CMS-11790 Use "required" validator to validate required fields
The hardcoded list of "unsupported field validators" no longer exists.
Instead, every validator name for which the ValidationService does not
return a ValidatorInstance is now called a unsupported validator. These
are the custom validators that only exist in Wicket; projects should
convert them to generic validators. We used to call those "unknown"
validators, but "unsupported" validators seemed a better term.