CMS-11839 Put navapp script sources in a FilteredHeaderItem
To prevent javascript error and library clashes (e.g. jquery) we want
separate the wicket + hippo cms css and js from those of the navapp.
This can be done by using a HeaderResponseDecorator. Such a decorator
can be used to put resources into buckets. As a developer you can then
decide where to render such a bucket in your body. You do that by
adding a HeaderResponseContainer to your container and html that refers
to the bucket. By wrapping the NavAppHeaderItem in a FilteredHeaderItem
it is easy to write a headre filter that creates a navapp-items bucket.
Luckily the FilteringHeaderResponse (which implements the
IHeaderResponseDecorator) has a protected render method that we can
override: if the page is rendered in "navApp" mode then we ignore all
non-navapp resources otherwise, just render items as usual.

Note that the filterName of the opposite filter does not really matter.
Wicket will put all of those in the header anyway. So as a convenience
I used the wicket default filter name.