CMS-11837 Hide the perspective menu.
Show the perspective menu only if there is a servlet filter init
parameter with name “showPerspectiveMenu” and value “true”, so that
there is a way to show the menu in case any client wants to still
use the perspective menu.

Overwrite the wireFrameSetting, so that the width of the “left”
position becomes 0. That setting is configured at
/hippo:configuration/hippo:frontend/cms/cms-static/root/layout.wireframe/left
id=tabbed-panel-layout-left,width=64. We do not bootstrap it with
width=0, because then option to show the perspective menu is lost.

Set the visibility of the User Menu and the AutoExportPlugin

For the footer item only add a left margin if the perspective menu
should be shown. This is done by adding a class with jquery in a
header item, because it footer item contains a child that is dynamically
added by the plugin framework. If we add a wicket:id to the footer
the extension mechanism fails. This seemed the simplest approach.

Move the logic of when to show the Navigation Application to a new service.