CMS-11890 Add iframe query parameter
The CMS must be loaded in an IFrame. Wicket will render the CMS in
"iframe" mode only if the query parameter named iframe is present.
The nav item base urls should therefore be suffixed with ?iframe so
that the nav app will render the CMS in an iframe indeed in iframe mode.