HSTTWO-4597 Support CM preview over live channel with acces token
If there is a valid access token, the HST should be able to return the
Channel Mgr preview response over the live channel (host of the live
channel instead of over the cms host for preview)

This is a POC only, which is not complete. The code needs to be improved
in multiple places, for example

1) The Token parsing : currently non signed tokens are assumed, while
we need signed tokens. Perhaps use Spring Security framework for this?
2) When an end project already implemented and relies on the Authorization
'Bearer' authentication, should we support a different Header?
3) The request processing is stateless now, without HttpSession. Hence
all code that relies on the SSO HttpSession should be modified to also
be able to work without HttpSession, for example the
CmsComponentWindowResponseAppender but also in downstream projects