CMS-11918 Only load vendor.js and styles.js in development mode
The navapp javascript distribution doe not contain the above javascript
resources. All of that code is contained in the main.js. So when you
do not run in development mode these script sources should not be
loaded. In addition, in non-development mode the style.css must get
loaded.

Note that I added a FIXME because wicket messes up the main.js contents
during compression!