CMS-11930 Use navapp communication lib to connect to parent iframe
At the time of coding the implementation of the ClientAPI part of the
navapp-communication the communication library was not available yet.
So the call for connecting to the parent was stubbed. Now that the
communication lib is available we replace the stub with a call to the
connectToParent function provided by the navapp communication library.