CMS-11942 (Re)Introduce a NavAppSettings factory
The NavAppHeaderItem was getting obese. I solved it by creating a
factory class that can create NavAppSettings instances.

Later, when brSM apps will be added we'll probably read configuration
from the repository. So then we can put that code in the factory too.

I also changed the location properties from URL to URI. I did this
because I want to compare two locations. Using equals on URLs has the
side-effect of resolving a host-name to an ip address. We don't need
that.