CMS-11942 Add antiCache query parameter if CMS serves the resources
When css and javascript resources are served by the cms then we must
make sure that they will get an antiCache query parameter. That way we
will know for sure that the resources won't be cached by the browser if
we release the CMS with a new version of the navapp resources.

We can make use