NAVAPP-30 Make home part of the left menu
The first item of the menu is designated to be the Home menu
item, irrespective of what it is. Therefore, the home menu item
should be in the menu.

Note that I added a filter to the navigationConfiguration so
that it will only emit non-empty maps. This way subscribers
do not have to take the special initial empty map into
account.