CMS-11947 use legacy validator mapper to support old style validator configuration
In the mock field type context the mock of the call to FieldTypeUtils#determineValidators was removed. It was not necessary for other tests and it plays a test role now in the new 'support legacy' tests.

Because getType is called now in the AbstractFieldType#init method the type is set for the default mocked fieldType in the test. The test for type is changed slightly but still tests what it should.

The 'support legacy' tests check if the legacy mapping results in the correct call to FieldTypeUtils#determineValidators.