NAVAPP-13 Add parent query parameter
To be able to setup a postMessage from an iframe to a parent we need to
know the parent url. This information should not come from a query
parameter, because that would allow any application to embed the child
iframe in their page. That is clearly a security issue. Instead, the
parent origin should be provided on the domain of the iframe itself.
This can e.g. be done via a cookie set by the server that serves the
source of the iframe.

Since we don't have that yet we use the unsafe query parameter for now.
That way we can already start testing the communication with apps hosted
outside of the CMS.