CHANNELMGR-2376 show/hide mask when modal dialog opens/closes
If a modal dialog opens in the channel manager then the user should not
be able to click on items in the navapp. Once the modal is closed the
menu items should be clickable again.

The global javascript object Hippo.AppToNavApp contains the functions
showMask and hideMask. Those functions will take care of communicating
the show/hide request via postMessage.