CMS-11946 Undetach JcrTypeLocator after using it
Somehow the system holds a reference to the JcrTypeStore inside the
locator when a violation is reported by the ContentBlocksValidator. As
a result, warnings are logged about undetached JcrItemModels for all
types inside the type store, except the type a violation is reported
for (probably because violations detach their own dependencies).
Detaching the type store once the ContentBlockValidator is done with it
ensures all internally created JcrItemModels are properly detached and
removes all warnings.