CMS-11969 Implement ClientApi.logout()
Use jQuery.ajax to invoke the logout behaviour. The method returns
an object that implements the Promise interface, so there is nothing
extra to do.

The benefit of calling the logout behavior is that it only performs a
server side logout without changing anything in the UI. This way the
login screen won't be displayed in the CMS iframe. However, there is a
small chance that the pinger detects the logout before the navapp
completes the logout, so then you would briefly see the login in the
iframe. Maybe we can stop the pinger somehow?

The logout behaviour must now always be added (in ActiveLogoutPlugin
constructor), even in development mode. To reduce the amount of argument
passing I decided to move the rendering of the navapp to app header item
from the root plugin to the active logout plugin. That's OK because the
root plugin creates an instance of the active logout plugin.