NAVAPP-60 Resolve/reject all promises
We must make sure that all promises returned from logout are
either resolved or rejected. Since Promise.all would stop on the
first rejection I wrap it in a promise that always resolves before
passing them to Promise.all. On calling site we can then inspect
the array for errors and do something with them. Logging them
is kind of useless though because the browser will refresh, so all
logs are lost.