CHANNELMGR-2361 Run compound validators for list-based content blocks
List-based content blocks are instantiated via a different code path
than provider-based content blocks, so they still need to read the
validators configured on the type of the blocks.

Probably all type-based validators will be read in the isCompound()
case. Still, customers could configure additional compound validators
on the built-in types (hippostd:html, hippo:mirror, etc), which should
be triggered too when they are used as content blocks. Those are covered
in the second case, where we instantiate a new FieldTypeContext for each
block.