NAVAPP-68 Respond to navigate and updateNavLocation
Split CommunicationsService into CommunicationsService and
AppToNavAppService to avoid circular dependencies ( The implementation
of the methods of the ParentApi needs a dependency on the
MenuStateService. That service has een dependency on the
CommunicationService ).

Rename MenuStateService#setActiveItem to
MenuStateService#setActiveItemAndNavigate and
add MenuStateService#setActiveItem to distinguish between only setting
the active item and setting the active item and navigate.