NAVAPP-52 Add siteResource to AppSettings
There should only be one resource that provides the sites needed for the
SM apps. Original plan was to iterate over all navApp IFRAME resources
to find the one that implemented ClientApi.getSites (there should only
be one) and then call that to get the Sites. However, that is a rather
brittle design. Instead we decided to make the sites resource explicit
in the App settings.