CHANNELMGR-2381 Reduce handle width, move animated border to handle
A resize handle in the left side panel of 5px wide overlaps a little too
much with the vertical scrollbar to my taste (at least in Chrome on
Linux). Changed the width to 3px instead.

Also "moved" the grey right border of the left side panel to the resize
handle. The border still looks the same but is now part of the resize
handle. The contents of the side panel can therefore move 1px to the
left, which saves another pixel of overlap on the vertical scrollbar:
the resize handle only overlaps 2px.

The border now becomes blue while hovered or dragged to make it more
clear that resizing is possible.