CMS-11978 Check if type is not a hippo:document before prepending path
Instead of only applying the path-prepend to fields of type COMPOUND, we
apply it to al node fields that are not a document.

Updated the failing unit test with a missing path; the unit test does
not correctly setup the ITypeDescriptor, hence it does not correctly
detect that this is not a document, and we need the content[1] path.