CMS-11990 Remove DOCTYPE from uploaded SVGs
Per 1.3. SVG namespace and DTD of the Scalable Vector Graphics (SVG)
2 W3C Working Draft 09 July 2015:

"A DTD is not provided in this specification, as the use of DTDs for
validating documents is known to be problematic. In particular, DTDs do
not handle namespaces gracefully and the range of constraints they can
express is limited. It is recommended that authors do not include a
DOCTYPE declaration in SVG documents."