TRIVIAL Suppress SonarQube warning about possible NPE
SonarQube analyzes that 'delegate' can be null here, since it's passed
to updateItem() which does a null-check on the parameter. However, when
delegate would be null, line 208 ("delegate.setPath(path)") would
already throw an NPE. Since this whole class assumes delegate is not
null, the warning can safely be suppressed.