TRIVIAL Re-enable validation of JcrTypeDescriptors in Wicket dev mode
In Wicket 6 the configuration type of a web application changed from
a string ("development") to an enum. The resulting apples-to-pears
comparison went unnoticed for several years until SonarQube reported it.
As a result the runtime validation of JcrTypeDescriptors has not been
done anymore for quite some time.

We should investigate any warnings logged before removing this runtime
check altogether.