ESSENTIALS-1267 add tooltip to the right of the link to warn users
If a user already has a CMS open in a tab and is logged in, opening another tab will not work nicely. One of the sessions will be logged out. This 'in your face' tooltip should make users very aware of this.
Some texts updated to work nicely with the text of the tooltip.