CHANNELMGR-2395 Use domain as parent origin
The supplied parent origin is used as targetorigin for html5 post
message [1] calls by penpall. Only if the origin of the parent matches
that of the parent origin messages are passed through.

Since the channel manager *always* has the same domain as the cms, we
use its own domain as parent origin.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Window/postMessage