REPO-1492 Add implicit read access for ancestry of jcr:path constraints
When using the hipposys:facet = jcr:path constraint, implicit read
access to the ancestry is given, thus not for hipposys:facet = jcr:uuid,
see https://documentation.bloomreach.com/library/concepts/security/domains.html

Implicit read access to the ancestry of the jcr:path constraint is
only given when :
1) org.hippoecm.repository.security.principals.FacetAuthPrincipal#getPrivileges
contains "jcr:read"
2) the sibling facet rules of the jcr:path facet rule do not exclude the
node that is pointed to by jcr:path
3) The implicit read access nodes are not automatically added to the
Lucene Authorization Query, meaning when doing a search, for example on
'node name', the search result might not include the implicit ancestry
nodes. In practice this will never be a problem. Supporting the implicit
reads in the authorization query is now too complex, though after
supporting REPO_2210 which does impact the authorization query
it might be doable

Implementation details: In the HippoAccessManager, I did not see the need
for 'initialized' boolean and checking for initialization. JCR Session
are not thread safe and should never be used concurrently. The
checkInitialized() method I do not see the purpose of, since Hippo
Repository (JR) fully controls the creation order of the Access Mngr and
the instance is tied to a JCR Session.