CHANNELMGR-2398 Remove breadcrumb
Remove breadcrumb as the functionality is now part of the navapp.

React to 'navigate-to-channel-overview' by showing the channel
overview.

Move updateNavLocation from _setChannel to makeRenderPath because
that methods always gets called when a channel is loaded.

Add channel id to path to inform the navapp of an updated path to
facilitate deeplinking.