NAVAPP-145 Show overview
Show overview when the root item of the breadcrumb is clicked.

Add '/overview' to path to distinguish between navigating to the
overview and navigating to the subpath the subapplication is in.

E.g. A user opens a channel ( subpath = channel.id ), then opens the
content perspective and opens the channel manager again. The channel
should still be open.