CMS-12086 Only notify navapp of activation menu items
Only notify navapp if a perspective has a corresponding menu item.

During startup the navigation application receieves a list of
menu items from the cms. That list contains plugins with a
config item "wicket:id=service.tab".

The EditPerspective is a perspective without menu item ( it's
part of the browser menu item ).