REPO-2215 Support not yet existing References in domain constraints
Instead of throwing exceptions when loading facet rules which reference
a non existing path, keep them with an empty value. Via
QFacetRule#referenceExists it is possible to check whether the reference
for the QFacetRule exists or not. Then depending on the 'equals' value,
the facet rule is either skipped *or* it results in the domain rule to
not match any nodes.

During the life time of a logged in session, a missing path reference
can be added. Then, the QFacetRule is updated (note that it's hash keeps
being the same since the hash is stored and the equals on instance
equality also still works). That the hash keeps being the same is handy
in case the QFacetRule is used in a HashMap/Set. Of course though tricky
to modify an object that is used in a HashMap as key or in a HashSet

Next to that, in the access manager, I needed the shared item statemanager
to retrieve paths, since using the present hierMgr resulted in recursion
when processing the added / removed nodes.