CMS-12096 Let navapp load its own resources
See navapp-64. let navapp load its own resources, so that when a
resource is added or its name is modified the cms does not have to be
changed ( apart from bumping the navapp version ).

As an additional benefit: cache bursting.

Postfix the navAppLocation with the jarPathPrefix so that if the
navapp is served from the cms (the navapp is packaged in jar ) the
resources are loaded from that location. That is the case
when the "navapp.location" property is empty.