NAVAPP-64 Use navAppResourceLocation as prefix
Use navAppResourceLocation instead of navAppLocation as prefix for
resources, because the navAppLocation is already in use for setting the
base href.
Add additional parameter navAppResourceLocation when the navapp is
served from a jar in the cms. In that case the resources are located
under "/navapp", the jarPathPrefix of the ResourceServlet defined for the
navapp resources.
See CMS-12096.