HSTTWO-4627 Fix the tests by cleaning attr value and use a cache
Instead of recomputing the user agent to be a browser or bot everytime,
cache the result

Since we detect more than just search engines, it is confusing to name
everything search engines. Hence instead, use 'search engine or bot'
as naming.

Also fixed the unit tests by clearing the cached value. As a result, now
one test fails since we have 'wget' and 'curl' defined as bots but they
are also in the list of typical browser user agents. We need to decide
whether we see wget and curl as normal web browser user agents or more
like 'bots'. I do think I prefer to mark them as 'bot' user agents and
modify the test expectation: For example for wget or curl I think we
for example never want to personalize the response, hence treat them as
if they are a crawler makes most sense to me