CMS-12043 Move context arrow 8px left on MacOS with hidden scrollbars
By default, MacOS renders scrollbars only when the user is scrolling,
and those scrollbars are overlayed so they don't take up space and don't
push other content to the side. Because of this, we need to push the
context-menu-link ourselves to prevent it from being rendered below the
scrollbar.

Also removed the Hippo.Set in favor of the native javascript Set and
simplified the impl of calculating the scrollbar width.