CMS-12119 Give read access to frontend config
The UI extensions separate domain config can be removed since it is
now covered by read access on /hippo:configuration/hippo:frontend

Note that it is in line with the changes in CHANNELMGR-2169 and
CMS-11872 because in there for the ui-extensions already read access to
everybody was given. Note that I think only read access to cms users
should be given. I'll create a separate issue for this