REPO-2224 Correct the 'edit' domain for documents
This new setup is different than in previous versions, BUT is more
precise.

With this configuration, you can only write to 'draft' documents if
the draft document either doesn't have the property hippostd:holder OR
if the hippostd:holder its value is equal to the current user.

This is the best ACL based setup since it disallows other users to
modify someones draft (unless you are admin of course)