CMS-12124 Do not update the nav location on perspective activation
Problem cause:
In the CMS the Perspective class calls updateNavLocation when the
perspective is activated, with a breadcrumb always set to null (since
there is no easy way to know the breadcrumb). When you click the channel
manager menu item the parentApi.updateNavLocation is called twice:
- once after the perspective is activated (with breadcrumb null).
- once when the (angular) channelMenuService has been instantiated. It
  will set the breadcrumb to the name of the current channel.

Due to postMessage the order in which these events are processed in the
parent (because the calls are converted to events) is undetermined. Most
of the time though, the event from the activated perspective (from the
CMS) is slower, at least in my browser. As a result, the breadcrumb gets
updated with a null value.

Problem solution:
Just don't call updateNavLocation after a perspective is activated.
There is no need to do so because the navapp will already set it to
the correct path.