CMS-11729 Use Scalr#crop instead of custom cropping algorithm
This allows us to scale only once instead of twice after the image is
cropped and removes the need for a custom scale algorithm as it will use
Scalr#resize. A second benefit of this is that cropping an image now
adheres to the configured "quality" setting of a variant.