REPO-2232 Refactor and harden the check permissions and privileges logic
A new (internal) PermissionManager singleton is introduced, together with a StandardPermissionNames constants class,
which standardizes handling, mapping and caching of all the standard JCR permission actions and privileges, as well as custom privileges.

The PermissionManager is used to retrofit all the current logic for parsing/mapping permission actions and/or privileges
in Session.checkPermission and the HippoAccessManager, including its related methods for the JCR AccessControlManager and
Jackrabbit AccessManager interfaces.
And no longer delegating to the default Jackrabbit implementation, removing the need to (always) catch a AccessDeniedException before proceeding.

The (Hippo) AccessManager created FacetAuthPrincipal class is enhanced to also provide the resolved privileges, with jcr:write and jcr:all replaced
with their aggregate standard privileges, to allow a more efficient and hardened implementation of the actual privileges checks done
in a new HippoAccessManager.hasPrivileges(Path path, Set<String> privilegeNames).

All usages of inline coded names of standard JCR privileges and default custom privileges are replaced with named constants in th PermissionNames constants class.