CMS-12148 Fix inter perspective navigation
If a call is made from the cms and the path is a prefix of the current
path then the updateNavLocation should not be called. E.g. First
the project perspective make a call to updateNavLocation with
{path:'projects/viwGPl/channels', breadcrumbLabel:'Easter Bunny'}
Then the Perspective make a call to updateNavLocation with
{path:'projects',breadcrumbLabel:'root'}
The second call should be ignored.

Revert the disablement of calling updateNavLocation upon Perspective
activation because in the case of links from one perspective to
another, e.g. the "Go to content" link the navapp should be informed.
Reverts commit d3a78495ea28940865d

Simplify the ParentApiCaller, because only updateNavLocation is used
and the breadcrumbLabel is always 'root' when it is called from the
Perspective.

Introduce a shared object 'currentNavLocation' in the Hippo namespace.
Sub-apps always set currentNavLocation before calling updateNavLocation.
The CMS only sets currentNavLocation if currentNavLocation.path does not
start with path.