REPO-2232 provide a fail fast check for hasPrivileges before trying to match per domain
When one or more privileges needs to be checked, first a containAll check is performed on the overall resolvedPrivileges for the user across all its domains.
This can especially speed up checks for special privileges, like hippo:admin, which only very few users will have.

To be able to do this fail fast check, a major cleanup and (internal impact only) refactoring has been done to replace the usage of FacetAuthPrincipals created
for each (matching) user domain, into a singleton principal per user session (stored in the subject).
The new/updated FacetAuthPrincipal now aggregates a set of (new) FacetAuthDomain objects, replacing the earlier usage of FacetAuthPrincipal.
And it aggregates all the FacetAuthDomain.getResolvedPrivileges() in a set of its own, which is used for the fail fast check referred to above.

As a side-effect (or because I was already adjusting the related code for the above purposes) also quite a lot of additional (internal impact only) cleanup has been done.