REPO-2221 Remove 'next fire time' if executed if there is no next fire time
First of all, the next fire time is already a system property meaning
it won't get overridden during bootstrap, so removing it is safe.

Next to that, if we keep the 'next fire time' for jobs that should
execute only once, it fires the next run again! Removing the property
solves this