REPO-2235 Mark runtime modifiable parts as 'system' instead of 'config'
Runtime, new jobs can be added below

/hippo:configuration/hippo:modules/scheduler/hippo:moduleconfig
The above node is bootstrapped by the repository, have it does not have

.meta:residual-child-node-category: system

meaning that on a new deploy when the 'hippo:moduleconfig' config has
changes, it replaces the entire 'hippo:moduleconfig'. This is
undesirable since it might remove runtime added jobs. The
hippo:moduleconfig should get '.meta:residual-child-node-category: system'
to avoid this

Bootstrapped configuration for the FormDataCleanup job and
EventLogCleanup should have a crontrigger that has the
'hipposched:enabled' property as system such that is switched off in
production, a new bootstrap will never reset it automatically.
'hipposched:enabled' should therefor be system