NAVAPP-219 Add support for INTERNAL_REST navConfig resources
The wicket application cannot generate absolute urls for REST resources
if it is behind a reverse proxy without introducing backwards
incompatible changes to virtual host configurations (that would cause a
lot of impact to customers and BRC).

To let the navapp call REST services we therefore introduce the resource
type INTERNAL_REST. If a resource is of that type then the navapp will
have to prefix the resource url with the navappRoot url, where the
navappRoot url is the "root" of the navapp, e.g. https://domain.com/foo.